options-infra/foundry-sk-agents/apps/sk-agents/requirements.txt
fastapi>=0.109.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0
pydantic>=2.0.0,<3.0.0
httpx>=0.27.0,<1.0.0

# Semantic Kernel with Azure support
semantic-kernel[azure]>=1.15.0

# Azure AI
azure-identity>=1.15.0

# OpenTelemetry - use azure-monitor-opentelemetry for auto-instrumentation
azure-monitor-opentelemetry>=1.6.0
opentelemetry-instrumentation-fastapi>=0.48b0
opentelemetry-instrumentation-httpx>=0.48b0
